backend/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
Flask-Cors==4.0.0
PyMongo==4.5.0  
pymysql==1.1.0  